SMLMMetrics

Documentation for SMLMMetrics.

SMLMMetrics.jaccard

SMLMMetrics.jaccard

SMLMMetrics.match

SMLMMetrics.rmse

SMLMMetrics.rmse

SMLMMetrics.jaccard — Method

jaccard(a,b,cutoff)

Calculate the Jaccard Index between the two sets a and b using a maximum connection distance of cutoff.

The Jaccard index is defined as:

$J(A,B) = \frac{|A\capB|}{|A\cupB|}$

The intersecting elements of $A$ and $B$ found by building and minimizing a cost matrix whose elements are:

$C_{i,j} = \begin{cases} ||A_i-B_j|| & \quad \text{if all } A_{i,k}-B_{j,k} < d_k \\infty & \quad \text{otherwise} \end{cases}$

SMLMMetrics.jaccard — Method

jaccard(a::Array{<:Real},b::Array{<:Real},cutoff::Vector[<:Real])

a and b are d x n, and d x m arrays, where d is the number of dimensions

cutoff contains the maximum matching distance for each dimension.

SMLMMetrics.match — Method

function match(a::Array(<:Real),b::Array(<:Real), cutooff::Vector{<:Real})

Find best match between points in a and b.

SMLMMetrics.rmse — Method

function rmse(a,b)

Calculate the root meen square error between a and b.

SMLMMetrics.rmse — Method

function rmse(a::Array{<:Real},b::Array{<:Real}, α::Vector{<:Real})

Calculate the root mean square error using a dimensional weighting α.

a and b are d x n, and d x m arrays, where d is the number of dimensions

$RMSE = (\frac{1}{d}\frac{1}{N}\Sigma_{k=1}^{d}\alpha_k^2\Sigma_{n=1}^N (a_{k,i}-b_{k,i})^2$)^{1/2}